removed a bunch of simulation files
Not needed for demo. Hopefully someone else will deal with this in the future.

+++ src/main/java/edu/illinois/mitra/cyphyhouse/models/Model_Quadcopter.java
@@ -17,7 +17,7 @@ import edu.illinois.mitra.cyphyhouse.objects.PositionList;
 public class Model_Quadcopter extends ItemPosition implements TrackedRobot{
 	// for default values, see initial_helper()
 
-	public double radius;
+	public int radius;
 	
 	public Random rand;
